requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
requests>=2.25.0
beautifulsoup4>=4.9.0
PyYAML>=5.4.0 